separate out test parser
